Map easy2see, find_next_gun and select_guns
object_collision calls 0x417610 twice: that is easy2see, sitting right
before make_shard.  It reads the screen-X entry three times rather than
caching it and passes the shape straight through, and the flip argument
goes out as a short -- every GetTransformedShapeBounds call site pushes
a 16-bit flip, and the callee ignores the parameter entirely, so the
prototype was simply wrong.  easy2see, DrawScaledTitleLogo and
select_new_gun now match exactly; draw_target_box and
build_your_target_list improve as a side effect.

select_new_gun's two helpers are 0x461480 and 0x4615A2.  select_guns
takes only the gun it is switching to -- there is no object argument --
and find_next_gun skips object type 11, the nameless projectile that
carries the laser cannon's stats and is what a turret fires.
select_new_gun itself returns void.

+++ src/gr.c
@@ -2189,7 +2189,7 @@ void DrawSolidColourSpriteScaled(Viewport *viewport, short x, short y,
 /* Function start: 0x4283A4 */
 int GetTransformedShapeBounds(Viewport *viewport, short x, short y,
                               unsigned char *shape, short frame,
-                              short angle, short scale, int flip,
+                              short angle, short scale, short flip,
                               short *bounds)
 {
     short *frameData;
